@@ -25,20 +25,21 @@
   },
   "devDependencies": {
     "browser-sync": "2.15.0",
-    "http-server": "0.9.0",
     "del": "2.1.0",
     "gulp": "^3.9.1",
     "gulp-clean": "0.3.2",
     "gulp-replace": "0.5.4",
-    "gulp-systemjs-builder": "^0.15.0",
+    "gulp-systemjs-builder": "0.15.0",
     "gulp-typescript": "^2.9.2",
     "gulp-watch": "^4.3.5",
+    "http-server": "0.9.0",
     "isarray": "1.0.0",
     "jspm": "0.16.53",
     "merge": "1.2.0",
-    "systemjs": "0.20.19",
-    "systemjs-builder": "0.16.12",
+    "rxjs-compat": "^6.2.0",
+    "systemjs": "0.19.0",
+    "systemjs-builder": "0.15.32",
     "traceur": "0.0.111",
-    "typescript": "2.4.2"
+    "typescript": "2.5.2"
   }
 }
