@@ -1,6 +1,6 @@
 import { Injectable }    from '@angular/core';
 import { CanDeactivate } from '@angular/router';
-import { Observable }    from 'rxjs/Observable';
+import { Observable }    from 'rxjs';
 
 export interface CanComponentDeactivate {
  canDeactivate: () => Observable<boolean> | Promise<boolean> | boolean;
