@@ -8,38 +8,50 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
 var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
-var core_1 = require('@angular/core');
-var router_1 = require('@angular/router');
-var selective_preload_strategy_1 = require('../selective-preload-strategy');
-require('rxjs/add/operator/map');
-var AdminDashboardComponent = (function () {
-    function AdminDashboardComponent(route, preloadStrategy) {
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const router_1 = require("@angular/router");
+const selective_preload_strategy_1 = require("../selective-preload-strategy");
+require("rxjs/add/operator/map");
+let AdminDashboardComponent = class AdminDashboardComponent {
+    constructor(route, preloadStrategy) {
         this.route = route;
         this.preloadStrategy = preloadStrategy;
         this.modules = preloadStrategy.preloadedModules;
     }
-    AdminDashboardComponent.prototype.ngOnInit = function () {
+    ngOnInit() {
         // Capture the session ID if available
         this.sessionId = this.route
             .queryParams
-            .map(function (params) { return params['session_id'] || 'None'; });
+            .map(params => params['session_id'] || 'None');
         // Capture the fragment if available
         this.token = this.route
             .fragment
-            .map(function (fragment) { return fragment || 'None'; });
-    };
-    AdminDashboardComponent = __decorate([
-        core_1.Component({
-            template: "\n    <p>Dashboard</p>\n\n    <p>Session ID: {{ sessionId | async }}</p>\n    <a id=\"anchor\"></a>\n    <p>Token: {{ token | async }}</p>\n\n    Preloaded Modules\n    <ul>\n      <li *ngFor=\"let module of modules\">{{ module }}</li>\n    </ul>\n  "
-        }), 
-        __metadata('design:paramtypes', [router_1.ActivatedRoute, selective_preload_strategy_1.PreloadSelectedModules])
-    ], AdminDashboardComponent);
-    return AdminDashboardComponent;
-}());
+            .map(fragment => fragment || 'None');
+    }
+};
+AdminDashboardComponent = __decorate([
+    core_1.Component({
+        template: `
+    <p>Dashboard</p>
+
+    <p>Session ID: {{ sessionId | async }}</p>
+    <a id="anchor"></a>
+    <p>Token: {{ token | async }}</p>
+
+    Preloaded Modules
+    <ul>
+      <li *ngFor="let module of modules">{{ module }}</li>
+    </ul>
+  `
+    }),
+    __metadata("design:paramtypes", [router_1.ActivatedRoute,
+        selective_preload_strategy_1.PreloadSelectedModules])
+], AdminDashboardComponent);
 exports.AdminDashboardComponent = AdminDashboardComponent;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
 Use of this source code is governed by an MIT-style license that
 can be found in the LICENSE file at http://angular.io/license
-*/ 
+*/
 //# sourceMappingURL=admin-dashboard.component.js.map
