@@ -5,21 +5,24 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+let AdminComponent = class AdminComponent {
 };
-var core_1 = require('@angular/core');
-var AdminComponent = (function () {
-    function AdminComponent() {
-    }
-    AdminComponent = __decorate([
-        core_1.Component({
-            template: "\n    <h3>ADMIN</h3>\n    <nav>\n      <a routerLink=\"./\" routerLinkActive=\"active\"\n        [routerLinkActiveOptions]=\"{ exact: true }\">Dashboard</a>\n      <a routerLink=\"./crises\" routerLinkActive=\"active\">Manage Crises</a>\n      <a routerLink=\"./heroes\" routerLinkActive=\"active\">Manage Heroes</a>\n    </nav>\n    <router-outlet></router-outlet>\n  "
-        }), 
-        __metadata('design:paramtypes', [])
-    ], AdminComponent);
-    return AdminComponent;
-}());
+AdminComponent = __decorate([
+    core_1.Component({
+        template: `
+    <h3>ADMIN</h3>
+    <nav>
+      <a routerLink="./" routerLinkActive="active"
+        [routerLinkActiveOptions]="{ exact: true }">Dashboard</a>
+      <a routerLink="./crises" routerLinkActive="active">Manage Crises</a>
+      <a routerLink="./heroes" routerLinkActive="active">Manage Heroes</a>
+    </nav>
+    <router-outlet></router-outlet>
+  `
+    })
+], AdminComponent);
 exports.AdminComponent = AdminComponent;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
