@@ -5,36 +5,30 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const common_1 = require("@angular/common");
+const admin_component_1 = require("./admin.component");
+const admin_dashboard_component_1 = require("./admin-dashboard.component");
+const manage_crises_component_1 = require("./manage-crises.component");
+const manage_heroes_component_1 = require("./manage-heroes.component");
+const admin_routing_module_1 = require("./admin-routing.module");
+let AdminModule = class AdminModule {
 };
-var core_1 = require('@angular/core');
-var common_1 = require('@angular/common');
-var admin_component_1 = require('./admin.component');
-var admin_dashboard_component_1 = require('./admin-dashboard.component');
-var manage_crises_component_1 = require('./manage-crises.component');
-var manage_heroes_component_1 = require('./manage-heroes.component');
-var admin_routing_module_1 = require('./admin-routing.module');
-var AdminModule = (function () {
-    function AdminModule() {
-    }
-    AdminModule = __decorate([
-        core_1.NgModule({
-            imports: [
-                common_1.CommonModule,
-                admin_routing_module_1.AdminRoutingModule
-            ],
-            declarations: [
-                admin_component_1.AdminComponent,
-                admin_dashboard_component_1.AdminDashboardComponent,
-                manage_crises_component_1.ManageCrisesComponent,
-                manage_heroes_component_1.ManageHeroesComponent
-            ]
-        }), 
-        __metadata('design:paramtypes', [])
-    ], AdminModule);
-    return AdminModule;
-}());
+AdminModule = __decorate([
+    core_1.NgModule({
+        imports: [
+            common_1.CommonModule,
+            admin_routing_module_1.AdminRoutingModule
+        ],
+        declarations: [
+            admin_component_1.AdminComponent,
+            admin_dashboard_component_1.AdminDashboardComponent,
+            manage_crises_component_1.ManageCrisesComponent,
+            manage_heroes_component_1.ManageHeroesComponent
+        ]
+    })
+], AdminModule);
 exports.AdminModule = AdminModule;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
