@@ -5,21 +5,17 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+let ManageCrisesComponent = class ManageCrisesComponent {
 };
-var core_1 = require('@angular/core');
-var ManageCrisesComponent = (function () {
-    function ManageCrisesComponent() {
-    }
-    ManageCrisesComponent = __decorate([
-        core_1.Component({
-            template: "\n    <p>Manage your crises here</p>\n  "
-        }), 
-        __metadata('design:paramtypes', [])
-    ], ManageCrisesComponent);
-    return ManageCrisesComponent;
-}());
+ManageCrisesComponent = __decorate([
+    core_1.Component({
+        template: `
+    <p>Manage your crises here</p>
+  `
+    })
+], ManageCrisesComponent);
 exports.ManageCrisesComponent = ManageCrisesComponent;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
