@@ -5,43 +5,37 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const platform_browser_1 = require("@angular/platform-browser");
+const forms_1 = require("@angular/forms");
+const app_component_1 = require("./app.component");
+const app_routing_module_1 = require("./app-routing.module");
+const heroes_module_1 = require("./heroes/heroes.module");
+const login_routing_module_1 = require("./login-routing.module");
+const login_component_1 = require("./login.component");
+const dialog_service_1 = require("./dialog.service");
+let AppModule = class AppModule {
 };
-var core_1 = require('@angular/core');
-var platform_browser_1 = require('@angular/platform-browser');
-var forms_1 = require('@angular/forms');
-var app_component_1 = require('./app.component');
-var app_routing_module_1 = require('./app-routing.module');
-var heroes_module_1 = require('./heroes/heroes.module');
-var login_routing_module_1 = require('./login-routing.module');
-var login_component_1 = require('./login.component');
-var dialog_service_1 = require('./dialog.service');
-var AppModule = (function () {
-    function AppModule() {
-    }
-    AppModule = __decorate([
-        core_1.NgModule({
-            imports: [
-                platform_browser_1.BrowserModule,
-                forms_1.FormsModule,
-                heroes_module_1.HeroesModule,
-                login_routing_module_1.LoginRoutingModule,
-                app_routing_module_1.AppRoutingModule
-            ],
-            declarations: [
-                app_component_1.AppComponent,
-                login_component_1.LoginComponent
-            ],
-            providers: [
-                dialog_service_1.DialogService
-            ],
-            bootstrap: [app_component_1.AppComponent]
-        }), 
-        __metadata('design:paramtypes', [])
-    ], AppModule);
-    return AppModule;
-}());
+AppModule = __decorate([
+    core_1.NgModule({
+        imports: [
+            platform_browser_1.BrowserModule,
+            forms_1.FormsModule,
+            heroes_module_1.HeroesModule,
+            login_routing_module_1.LoginRoutingModule,
+            app_routing_module_1.AppRoutingModule
+        ],
+        declarations: [
+            app_component_1.AppComponent,
+            login_component_1.LoginComponent
+        ],
+        providers: [
+            dialog_service_1.DialogService
+        ],
+        bootstrap: [app_component_1.AppComponent]
+    })
+], AppModule);
 exports.AppModule = AppModule;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
