@@ -8,49 +8,49 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
 var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
-var core_1 = require('@angular/core');
-var router_1 = require('@angular/router');
-var auth_service_1 = require('./auth.service');
-var AuthGuard = (function () {
-    function AuthGuard(authService, router) {
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const router_1 = require("@angular/router");
+const auth_service_1 = require("./auth.service");
+let AuthGuard = class AuthGuard {
+    constructor(authService, router) {
         this.authService = authService;
         this.router = router;
     }
-    AuthGuard.prototype.canActivate = function (route, state) {
-        var url = state.url;
+    canActivate(route, state) {
+        let url = state.url;
         return this.checkLogin(url);
-    };
-    AuthGuard.prototype.canActivateChild = function (route, state) {
+    }
+    canActivateChild(route, state) {
         return this.canActivate(route, state);
-    };
-    AuthGuard.prototype.canLoad = function (route) {
-        var url = "/" + route.path;
+    }
+    canLoad(route) {
+        let url = `/${route.path}`;
         return this.checkLogin(url);
-    };
-    AuthGuard.prototype.checkLogin = function (url) {
+    }
+    checkLogin(url) {
         if (this.authService.isLoggedIn) {
             return true;
         }
         // Store the attempted URL for redirecting
         this.authService.redirectUrl = url;
         // Create a dummy session id
-        var sessionId = 123456789;
+        let sessionId = 123456789;
         // Set our navigation extras object
         // that contains our global query params and fragment
-        var navigationExtras = {
+        let navigationExtras = {
             queryParams: { 'session_id': sessionId },
             fragment: 'anchor'
         };
         // Navigate to the login page with extras
         this.router.navigate(['/login'], navigationExtras);
         return false;
-    };
-    AuthGuard = __decorate([
-        core_1.Injectable(), 
-        __metadata('design:paramtypes', [auth_service_1.AuthService, router_1.Router])
-    ], AuthGuard);
-    return AuthGuard;
-}());
+    }
+};
+AuthGuard = __decorate([
+    core_1.Injectable(),
+    __metadata("design:paramtypes", [auth_service_1.AuthService, router_1.Router])
+], AuthGuard);
 exports.AuthGuard = AuthGuard;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
