@@ -5,35 +5,28 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
-};
-var core_1 = require('@angular/core');
-var Observable_1 = require('rxjs/Observable');
-require('rxjs/add/observable/of');
-require('rxjs/add/operator/do');
-require('rxjs/add/operator/delay');
-var AuthService = (function () {
-    function AuthService() {
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const of_1 = require("rxjs/observable/of");
+const tap_1 = require("rxjs/operators/tap");
+let AuthService = class AuthService {
+    constructor() {
         this.isLoggedIn = false;
     }
-    AuthService.prototype.login = function () {
-        var _this = this;
-        return Observable_1.Observable.of(true).delay(1000).do(function (val) { return _this.isLoggedIn = true; });
-    };
-    AuthService.prototype.logout = function () {
+    login() {
+        return of_1.of(true).delay(1000).pipe(tap_1.tap(val => this.isLoggedIn = true));
+    }
+    logout() {
         this.isLoggedIn = false;
-    };
-    AuthService = __decorate([
-        core_1.Injectable(), 
-        __metadata('design:paramtypes', [])
-    ], AuthService);
-    return AuthService;
-}());
+    }
+};
+AuthService = __decorate([
+    core_1.Injectable()
+], AuthService);
 exports.AuthService = AuthService;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
 Use of this source code is governed by an MIT-style license that
 can be found in the LICENSE file at http://angular.io/license
-*/ 
+*/
 //# sourceMappingURL=auth.service.js.map
