@@ -5,26 +5,20 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
-};
-var core_1 = require('@angular/core');
-var CanDeactivateGuard = (function () {
-    function CanDeactivateGuard() {
-    }
-    CanDeactivateGuard.prototype.canDeactivate = function (component) {
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+let CanDeactivateGuard = class CanDeactivateGuard {
+    canDeactivate(component) {
         return component.canDeactivate ? component.canDeactivate() : true;
-    };
-    CanDeactivateGuard = __decorate([
-        core_1.Injectable(), 
-        __metadata('design:paramtypes', [])
-    ], CanDeactivateGuard);
-    return CanDeactivateGuard;
-}());
+    }
+};
+CanDeactivateGuard = __decorate([
+    core_1.Injectable()
+], CanDeactivateGuard);
 exports.CanDeactivateGuard = CanDeactivateGuard;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
 Use of this source code is governed by an MIT-style license that
 can be found in the LICENSE file at http://angular.io/license
-*/ 
+*/
 //# sourceMappingURL=can-deactivate-guard.service.js.map
