@@ -5,61 +5,55 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const router_1 = require("@angular/router");
+const crisis_center_home_component_1 = require("./crisis-center-home.component");
+const crisis_list_component_1 = require("./crisis-list.component");
+const crisis_center_component_1 = require("./crisis-center.component");
+const crisis_detail_component_1 = require("./crisis-detail.component");
+const can_deactivate_guard_service_1 = require("../can-deactivate-guard.service");
+const crisis_detail_resolve_service_1 = require("./crisis-detail-resolve.service");
+let CrisisCenterRoutingModule = class CrisisCenterRoutingModule {
 };
-var core_1 = require('@angular/core');
-var router_1 = require('@angular/router');
-var crisis_center_home_component_1 = require('./crisis-center-home.component');
-var crisis_list_component_1 = require('./crisis-list.component');
-var crisis_center_component_1 = require('./crisis-center.component');
-var crisis_detail_component_1 = require('./crisis-detail.component');
-var can_deactivate_guard_service_1 = require('../can-deactivate-guard.service');
-var crisis_detail_resolve_service_1 = require('./crisis-detail-resolve.service');
-var CrisisCenterRoutingModule = (function () {
-    function CrisisCenterRoutingModule() {
-    }
-    CrisisCenterRoutingModule = __decorate([
-        core_1.NgModule({
-            imports: [
-                router_1.RouterModule.forChild([
-                    {
-                        path: '',
-                        component: crisis_center_component_1.CrisisCenterComponent,
-                        children: [
-                            {
-                                path: '',
-                                component: crisis_list_component_1.CrisisListComponent,
-                                children: [
-                                    {
-                                        path: ':id',
-                                        component: crisis_detail_component_1.CrisisDetailComponent,
-                                        canDeactivate: [can_deactivate_guard_service_1.CanDeactivateGuard],
-                                        resolve: {
-                                            crisis: crisis_detail_resolve_service_1.CrisisDetailResolve
-                                        }
-                                    },
-                                    {
-                                        path: '',
-                                        component: crisis_center_home_component_1.CrisisCenterHomeComponent
+CrisisCenterRoutingModule = __decorate([
+    core_1.NgModule({
+        imports: [
+            router_1.RouterModule.forChild([
+                {
+                    path: '',
+                    component: crisis_center_component_1.CrisisCenterComponent,
+                    children: [
+                        {
+                            path: '',
+                            component: crisis_list_component_1.CrisisListComponent,
+                            children: [
+                                {
+                                    path: ':id',
+                                    component: crisis_detail_component_1.CrisisDetailComponent,
+                                    canDeactivate: [can_deactivate_guard_service_1.CanDeactivateGuard],
+                                    resolve: {
+                                        crisis: crisis_detail_resolve_service_1.CrisisDetailResolve
                                     }
-                                ]
-                            }
-                        ]
-                    }
-                ])
-            ],
-            exports: [
-                router_1.RouterModule
-            ],
-            providers: [
-                crisis_detail_resolve_service_1.CrisisDetailResolve
-            ]
-        }), 
-        __metadata('design:paramtypes', [])
-    ], CrisisCenterRoutingModule);
-    return CrisisCenterRoutingModule;
-}());
+                                },
+                                {
+                                    path: '',
+                                    component: crisis_center_home_component_1.CrisisCenterHomeComponent
+                                }
+                            ]
+                        }
+                    ]
+                }
+            ])
+        ],
+        exports: [
+            router_1.RouterModule
+        ],
+        providers: [
+            crisis_detail_resolve_service_1.CrisisDetailResolve
+        ]
+    })
+], CrisisCenterRoutingModule);
 exports.CrisisCenterRoutingModule = CrisisCenterRoutingModule;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
