@@ -8,51 +8,39 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
 var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
-var core_1 = require('@angular/core');
-var router_1 = require('@angular/router');
-var dialog_service_1 = require('../dialog.service');
-var CrisisDetailComponent = (function () {
-    function CrisisDetailComponent(route, router, dialogService) {
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const router_1 = require("@angular/router");
+const dialog_service_1 = require("../dialog.service");
+let CrisisDetailComponent = class CrisisDetailComponent {
+    constructor(route, router, dialogService) {
         this.route = route;
         this.router = router;
         this.dialogService = dialogService;
     }
-    Object.defineProperty(CrisisDetailComponent.prototype, "routeAnimation", {
-        get: function () {
-            return true;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    Object.defineProperty(CrisisDetailComponent.prototype, "display", {
-        get: function () {
-            return 'block';
-        },
-        enumerable: true,
-        configurable: true
-    });
-    Object.defineProperty(CrisisDetailComponent.prototype, "position", {
-        get: function () {
-            return 'absolute';
-        },
-        enumerable: true,
-        configurable: true
-    });
-    CrisisDetailComponent.prototype.ngOnInit = function () {
-        var _this = this;
-        this.route.data.forEach(function (data) {
-            _this.editName = data.crisis.name;
-            _this.crisis = data.crisis;
+    get routeAnimation() {
+        return true;
+    }
+    get display() {
+        return 'block';
+    }
+    get position() {
+        return 'absolute';
+    }
+    ngOnInit() {
+        this.route.data.forEach((data) => {
+            this.editName = data.crisis.name;
+            this.crisis = data.crisis;
         });
-    };
-    CrisisDetailComponent.prototype.cancel = function () {
+    }
+    cancel() {
         this.gotoCrises();
-    };
-    CrisisDetailComponent.prototype.save = function () {
+    }
+    save() {
         this.crisis.name = this.editName;
         this.gotoCrises();
-    };
-    CrisisDetailComponent.prototype.canDeactivate = function () {
+    }
+    canDeactivate() {
         // Allow synchronous navigation (`true`) if no crisis or the crisis is unchanged
         if (!this.crisis || this.crisis.name === this.editName) {
             return true;
@@ -60,57 +48,75 @@ var CrisisDetailComponent = (function () {
         // Otherwise ask the user with the dialog service and return its
         // promise which resolves to true or false when the user decides
         return this.dialogService.confirm('Discard changes?');
-    };
-    CrisisDetailComponent.prototype.gotoCrises = function () {
-        var crisisId = this.crisis ? this.crisis.id : null;
+    }
+    gotoCrises() {
+        let crisisId = this.crisis ? this.crisis.id : null;
         // Pass along the crisis id if available
         // so that the CrisisListComponent can select that crisis.
         // Add a totally useless `foo` parameter for kicks.
         // Relative navigation back to the crises
         this.router.navigate(['../', { id: crisisId, foo: 'foo' }], { relativeTo: this.route });
-    };
-    __decorate([
-        core_1.HostBinding('@routeAnimation'), 
-        __metadata('design:type', Object)
-    ], CrisisDetailComponent.prototype, "routeAnimation", null);
-    __decorate([
-        core_1.HostBinding('style.display'), 
-        __metadata('design:type', Object)
-    ], CrisisDetailComponent.prototype, "display", null);
-    __decorate([
-        core_1.HostBinding('style.position'), 
-        __metadata('design:type', Object)
-    ], CrisisDetailComponent.prototype, "position", null);
-    CrisisDetailComponent = __decorate([
-        core_1.Component({
-            template: "\n  <div *ngIf=\"crisis\">\n    <h3>\"{{editName}}\"</h3>\n    <div>\n      <label>Id: </label>{{crisis.id}}</div>\n    <div>\n      <label>Name: </label>\n      <input [(ngModel)]=\"editName\" placeholder=\"name\"/>\n    </div>\n    <p>\n      <button (click)=\"save()\">Save</button>\n      <button (click)=\"cancel()\">Cancel</button>\n    </p>\n  </div>\n  ",
-            styles: ['input {width: 20em}'],
-            animations: [
-                core_1.trigger('routeAnimation', [
-                    core_1.state('*', core_1.style({
-                        opacity: 1,
-                        transform: 'translateX(0)'
-                    })),
-                    core_1.transition(':enter', [
-                        core_1.style({
-                            opacity: 0,
-                            transform: 'translateX(-100%)'
-                        }),
-                        core_1.animate('0.2s ease-in')
-                    ]),
-                    core_1.transition(':leave', [
-                        core_1.animate('0.5s ease-out', core_1.style({
-                            opacity: 0,
-                            transform: 'translateY(100%)'
-                        }))
-                    ])
+    }
+};
+__decorate([
+    core_1.HostBinding('@routeAnimation'),
+    __metadata("design:type", Object),
+    __metadata("design:paramtypes", [])
+], CrisisDetailComponent.prototype, "routeAnimation", null);
+__decorate([
+    core_1.HostBinding('style.display'),
+    __metadata("design:type", Object),
+    __metadata("design:paramtypes", [])
+], CrisisDetailComponent.prototype, "display", null);
+__decorate([
+    core_1.HostBinding('style.position'),
+    __metadata("design:type", Object),
+    __metadata("design:paramtypes", [])
+], CrisisDetailComponent.prototype, "position", null);
+CrisisDetailComponent = __decorate([
+    core_1.Component({
+        template: `
+  <div *ngIf="crisis">
+    <h3>"{{editName}}"</h3>
+    <div>
+      <label>Id: </label>{{crisis.id}}</div>
+    <div>
+      <label>Name: </label>
+      <input [(ngModel)]="editName" placeholder="name"/>
+    </div>
+    <p>
+      <button (click)="save()">Save</button>
+      <button (click)="cancel()">Cancel</button>
+    </p>
+  </div>
+  `,
+        styles: ['input {width: 20em}'],
+        animations: [
+            core_1.trigger('routeAnimation', [
+                core_1.state('*', core_1.style({
+                    opacity: 1,
+                    transform: 'translateX(0)'
+                })),
+                core_1.transition(':enter', [
+                    core_1.style({
+                        opacity: 0,
+                        transform: 'translateX(-100%)'
+                    }),
+                    core_1.animate('0.2s ease-in')
+                ]),
+                core_1.transition(':leave', [
+                    core_1.animate('0.5s ease-out', core_1.style({
+                        opacity: 0,
+                        transform: 'translateY(100%)'
+                    }))
                 ])
-            ]
-        }), 
-        __metadata('design:paramtypes', [router_1.ActivatedRoute, router_1.Router, dialog_service_1.DialogService])
-    ], CrisisDetailComponent);
-    return CrisisDetailComponent;
-}());
+            ])
+        ]
+    }),
+    __metadata("design:paramtypes", [router_1.ActivatedRoute,
+        router_1.Router,
+        dialog_service_1.DialogService])
+], CrisisDetailComponent);
 exports.CrisisDetailComponent = CrisisDetailComponent;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
