@@ -5,48 +5,42 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
-};
-var Crisis = (function () {
-    function Crisis(id, name) {
+Object.defineProperty(exports, "__esModule", { value: true });
+class Crisis {
+    constructor(id, name) {
         this.id = id;
         this.name = name;
     }
-    return Crisis;
-}());
+}
 exports.Crisis = Crisis;
-var CRISES = [
+const CRISES = [
     new Crisis(1, 'Dragon Burning Cities'),
     new Crisis(2, 'Sky Rains Great White Sharks'),
     new Crisis(3, 'Giant Asteroid Heading For Earth'),
     new Crisis(4, 'Procrastinators Meeting Delayed Again'),
 ];
-var crisesPromise = Promise.resolve(CRISES);
-var core_1 = require('@angular/core');
-var CrisisService = (function () {
-    function CrisisService() {
-    }
-    CrisisService.prototype.getCrises = function () { return crisesPromise; };
-    CrisisService.prototype.getCrisis = function (id) {
+let crisesPromise = Promise.resolve(CRISES);
+const core_1 = require("@angular/core");
+let CrisisService = CrisisService_1 = class CrisisService {
+    getCrises() { return crisesPromise; }
+    getCrisis(id) {
         return crisesPromise
-            .then(function (crises) { return crises.find(function (crisis) { return crisis.id === +id; }); });
-    };
-    CrisisService.prototype.addCrisis = function (name) {
+            .then(crises => crises.find(crisis => crisis.id === +id));
+    }
+    addCrisis(name) {
         name = name.trim();
         if (name) {
-            var crisis_1 = new Crisis(CrisisService.nextCrisisId++, name);
-            crisesPromise.then(function (crises) { return crises.push(crisis_1); });
+            let crisis = new Crisis(CrisisService_1.nextCrisisId++, name);
+            crisesPromise.then(crises => crises.push(crisis));
         }
-    };
-    CrisisService.nextCrisisId = 100;
-    CrisisService = __decorate([
-        core_1.Injectable(), 
-        __metadata('design:paramtypes', [])
-    ], CrisisService);
-    return CrisisService;
-}());
+    }
+};
+CrisisService.nextCrisisId = 100;
+CrisisService = CrisisService_1 = __decorate([
+    core_1.Injectable()
+], CrisisService);
 exports.CrisisService = CrisisService;
+var CrisisService_1;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
 Use of this source code is governed by an MIT-style license that
