@@ -8,90 +8,96 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
 var __metadata = (this && this.__metadata) || function (k, v) {
     if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
 };
-var core_1 = require('@angular/core');
-var router_1 = require('@angular/router');
-var hero_service_1 = require('./hero.service');
-var HeroDetailComponent = (function () {
-    function HeroDetailComponent(route, router, service) {
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const router_1 = require("@angular/router");
+const hero_service_1 = require("./hero.service");
+let HeroDetailComponent = class HeroDetailComponent {
+    constructor(route, router, service) {
         this.route = route;
         this.router = router;
         this.service = service;
     }
-    Object.defineProperty(HeroDetailComponent.prototype, "routeAnimation", {
-        get: function () {
-            return true;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    Object.defineProperty(HeroDetailComponent.prototype, "display", {
-        get: function () {
-            return 'block';
-        },
-        enumerable: true,
-        configurable: true
-    });
-    Object.defineProperty(HeroDetailComponent.prototype, "position", {
-        get: function () {
-            return 'absolute';
-        },
-        enumerable: true,
-        configurable: true
-    });
-    HeroDetailComponent.prototype.ngOnInit = function () {
-        var _this = this;
-        this.route.params.forEach(function (params) {
-            var id = +params['id']; // (+) converts string 'id' to a number
-            _this.service.getHero(id).then(function (hero) { return _this.hero = hero; });
+    get routeAnimation() {
+        return true;
+    }
+    get display() {
+        return 'block';
+    }
+    get position() {
+        return 'absolute';
+    }
+    ngOnInit() {
+        this.route.params.forEach((params) => {
+            let id = +params['id']; // (+) converts string 'id' to a number
+            this.service.getHero(id).then(hero => this.hero = hero);
         });
-    };
-    HeroDetailComponent.prototype.gotoHeroes = function () {
-        var heroId = this.hero ? this.hero.id : null;
+    }
+    gotoHeroes() {
+        let heroId = this.hero ? this.hero.id : null;
         // Pass along the hero id if available
         // so that the HeroList component can select that hero.
         this.router.navigate(['/heroes', { id: heroId, foo: 'foo' }]);
-    };
-    __decorate([
-        core_1.HostBinding('@routeAnimation'), 
-        __metadata('design:type', Object)
-    ], HeroDetailComponent.prototype, "routeAnimation", null);
-    __decorate([
-        core_1.HostBinding('style.display'), 
-        __metadata('design:type', Object)
-    ], HeroDetailComponent.prototype, "display", null);
-    __decorate([
-        core_1.HostBinding('style.position'), 
-        __metadata('design:type', Object)
-    ], HeroDetailComponent.prototype, "position", null);
-    HeroDetailComponent = __decorate([
-        core_1.Component({
-            template: "\n  <h2>HEROES</h2>\n  <div *ngIf=\"hero\">\n    <h3>\"{{hero.name}}\"</h3>\n    <div>\n      <label>Id: </label>{{hero.id}}</div>\n    <div>\n      <label>Name: </label>\n      <input [(ngModel)]=\"hero.name\" placeholder=\"name\"/>\n    </div>\n    <p>\n      <button (click)=\"gotoHeroes()\">Back</button>\n    </p>\n  </div>\n  ",
-            animations: [
-                core_1.trigger('routeAnimation', [
-                    core_1.state('*', core_1.style({
-                        opacity: 1,
-                        transform: 'translateX(0)'
-                    })),
-                    core_1.transition(':enter', [
-                        core_1.style({
-                            opacity: 0,
-                            transform: 'translateX(-100%)'
-                        }),
-                        core_1.animate('0.2s ease-in')
-                    ]),
-                    core_1.transition(':leave', [
-                        core_1.animate('0.5s ease-out', core_1.style({
-                            opacity: 0,
-                            transform: 'translateY(100%)'
-                        }))
-                    ])
+    }
+};
+__decorate([
+    core_1.HostBinding('@routeAnimation'),
+    __metadata("design:type", Object),
+    __metadata("design:paramtypes", [])
+], HeroDetailComponent.prototype, "routeAnimation", null);
+__decorate([
+    core_1.HostBinding('style.display'),
+    __metadata("design:type", Object),
+    __metadata("design:paramtypes", [])
+], HeroDetailComponent.prototype, "display", null);
+__decorate([
+    core_1.HostBinding('style.position'),
+    __metadata("design:type", Object),
+    __metadata("design:paramtypes", [])
+], HeroDetailComponent.prototype, "position", null);
+HeroDetailComponent = __decorate([
+    core_1.Component({
+        template: `
+  <h2>HEROES</h2>
+  <div *ngIf="hero">
+    <h3>"{{hero.name}}"</h3>
+    <div>
+      <label>Id: </label>{{hero.id}}</div>
+    <div>
+      <label>Name: </label>
+      <input [(ngModel)]="hero.name" placeholder="name"/>
+    </div>
+    <p>
+      <button (click)="gotoHeroes()">Back</button>
+    </p>
+  </div>
+  `,
+        animations: [
+            core_1.trigger('routeAnimation', [
+                core_1.state('*', core_1.style({
+                    opacity: 1,
+                    transform: 'translateX(0)'
+                })),
+                core_1.transition(':enter', [
+                    core_1.style({
+                        opacity: 0,
+                        transform: 'translateX(-100%)'
+                    }),
+                    core_1.animate('0.2s ease-in')
+                ]),
+                core_1.transition(':leave', [
+                    core_1.animate('0.5s ease-out', core_1.style({
+                        opacity: 0,
+                        transform: 'translateY(100%)'
+                    }))
                 ])
-            ]
-        }), 
-        __metadata('design:paramtypes', [router_1.ActivatedRoute, router_1.Router, hero_service_1.HeroService])
-    ], HeroDetailComponent);
-    return HeroDetailComponent;
-}());
+            ])
+        ]
+    }),
+    __metadata("design:paramtypes", [router_1.ActivatedRoute,
+        router_1.Router,
+        hero_service_1.HeroService])
+], HeroDetailComponent);
 exports.HeroDetailComponent = HeroDetailComponent;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
