@@ -5,19 +5,16 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
-};
-var core_1 = require('@angular/core');
-var Hero = (function () {
-    function Hero(id, name) {
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+class Hero {
+    constructor(id, name) {
         this.id = id;
         this.name = name;
     }
-    return Hero;
-}());
+}
 exports.Hero = Hero;
-var HEROES = [
+let HEROES = [
     new Hero(11, 'Mr. Nice'),
     new Hero(12, 'Narco'),
     new Hero(13, 'Bombasto'),
@@ -25,21 +22,17 @@ var HEROES = [
     new Hero(15, 'Magneta'),
     new Hero(16, 'RubberMan')
 ];
-var heroesPromise = Promise.resolve(HEROES);
-var HeroService = (function () {
-    function HeroService() {
-    }
-    HeroService.prototype.getHeroes = function () { return heroesPromise; };
-    HeroService.prototype.getHero = function (id) {
+let heroesPromise = Promise.resolve(HEROES);
+let HeroService = class HeroService {
+    getHeroes() { return heroesPromise; }
+    getHero(id) {
         return heroesPromise
-            .then(function (heroes) { return heroes.find(function (hero) { return hero.id === +id; }); });
-    };
-    HeroService = __decorate([
-        core_1.Injectable(), 
-        __metadata('design:paramtypes', [])
-    ], HeroService);
-    return HeroService;
-}());
+            .then(heroes => heroes.find(hero => hero.id === +id));
+    }
+};
+HeroService = __decorate([
+    core_1.Injectable()
+], HeroService);
 exports.HeroService = HeroService;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
