@@ -5,32 +5,26 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
+Object.defineProperty(exports, "__esModule", { value: true });
+const core_1 = require("@angular/core");
+const router_1 = require("@angular/router");
+const hero_list_component_1 = require("./hero-list.component");
+const hero_detail_component_1 = require("./hero-detail.component");
+let HeroRoutingModule = class HeroRoutingModule {
 };
-var core_1 = require('@angular/core');
-var router_1 = require('@angular/router');
-var hero_list_component_1 = require('./hero-list.component');
-var hero_detail_component_1 = require('./hero-detail.component');
-var HeroRoutingModule = (function () {
-    function HeroRoutingModule() {
-    }
-    HeroRoutingModule = __decorate([
-        core_1.NgModule({
-            imports: [
-                router_1.RouterModule.forChild([
-                    { path: 'heroes', component: hero_list_component_1.HeroListComponent },
-                    { path: 'hero/:id', component: hero_detail_component_1.HeroDetailComponent }
-                ])
-            ],
-            exports: [
-                router_1.RouterModule
-            ]
-        }), 
-        __metadata('design:paramtypes', [])
-    ], HeroRoutingModule);
-    return HeroRoutingModule;
-}());
+HeroRoutingModule = __decorate([
+    core_1.NgModule({
+        imports: [
+            router_1.RouterModule.forChild([
+                { path: 'heroes', component: hero_list_component_1.HeroListComponent },
+                { path: 'hero/:id', component: hero_detail_component_1.HeroDetailComponent }
+            ])
+        ],
+        exports: [
+            router_1.RouterModule
+        ]
+    })
+], HeroRoutingModule);
 exports.HeroRoutingModule = HeroRoutingModule;
 /*
 Copyright 2016 Google Inc. All Rights Reserved.
