@@ -1,27 +1,18 @@
-/**
- * WEB ANGULAR VERSION
- * (based on systemjs.config.js in angular.io)
- * System configuration for Angular samples
- * Adjust as necessary for your application needs.
- */
 (function (global) {
+  var angVer = '5.0.0';
+
   System.config({
-    // DEMO ONLY! REAL CODE SHOULD NOT TRANSPILE IN THE BROWSER
     transpiler: 'ts',
     typescriptOptions: {
-      // Complete copy of compiler options in standard tsconfig.json
-      "target": "es5",
-      "module": "commonjs",
-      "moduleResolution": "node",
-      "sourceMap": true,
       "emitDecoratorMetadata": true,
       "experimentalDecorators": true,
-      "removeComments": false,
+      "lib": [ "es2015", "dom" ],
+      "module": "commonjs",
+      "moduleResolution": "node",
       "noImplicitAny": true,
+      "sourceMap": true,
       "suppressImplicitAnyIndexErrors": true,
-      "typeRoots": [
-        "../../node_modules/@types/"
-      ]
+      "target": "es5"
     },
     meta: {
       'typescript': {
@@ -29,35 +20,29 @@
       }
     },
     paths: {
-      // paths serve as alias
-      'npm:': 'https://unpkg.com/'
+      'npm:': 'https://unpkg.com/',
+      'gh:': 'https://raw.githubusercontent.com/'
     },
-    // map tells the System loader where to look for things
     map: {
-      // our app is within the app folder
-      app: 'bundled',
-
-      // angular bundles
-      '@angular/core': 'npm:@angular/core/bundles/core.umd.js',
-      '@angular/common': 'npm:@angular/common/bundles/common.umd.js',
-      '@angular/compiler': 'npm:@angular/compiler/bundles/compiler.umd.js',
-      '@angular/platform-browser': 'npm:@angular/platform-browser/bundles/platform-browser.umd.js',
-      '@angular/platform-browser-dynamic': 'npm:@angular/platform-browser-dynamic/bundles/platform-browser-dynamic.umd.js',
-      '@angular/http': 'npm:@angular/http/bundles/http.umd.js',
-      '@angular/router': 'npm:@angular/router/bundles/router.umd.js',
-      '@angular/forms': 'npm:@angular/forms/bundles/forms.umd.js',
-      '@angular/upgrade': 'npm:@angular/upgrade/bundles/upgrade.umd.js',
-
-      // other libraries
-      'rxjs':                      'npm:rxjs',
-      'angular-in-memory-web-api': 'npm:angular-in-memory-web-api/bundles/in-memory-web-api.umd.js',
-      'ts':                        'npm:plugin-typescript@4.0.10/lib/plugin.js',
-      'typescript':                'npm:typescript@2.0.3/lib/typescript.js',
-      
-    },
+      "app": "bundled",
+      "rxjs": "npm:rxjs@5.5",
+      "ts": "npm:plugin-typescript@7.1.0/lib/plugin.js",
+      "typescript": "npm:typescript@2.4.2/lib/typescript.js",
+      "@angular/core": "npm:@angular/core@" + angVer + "/bundles/core.umd.js",
+      "@angular/common": "npm:@angular/common@" + angVer + "/bundles/common.umd.js",
+      "@angular/compiler": "npm:@angular/compiler@" + angVer + "/bundles/compiler.umd.js",
+      "@angular/platform-browser": "npm:@angular/platform-browser@" + angVer + "/bundles/platform-browser.umd.js",
+      "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic@" + angVer + "/bundles/platform-browser-dynamic.umd.js",
+      "@angular/http": "npm:@angular/http@" + angVer + "/bundles/http.umd.js",
+      "@angular/router": "npm:@angular/router@" + angVer + "/bundles/router.umd.js",
+      "@angular/forms": "npm:@angular/forms@" + angVer + "/bundles/forms.umd.js",
+      "@angular/animations": "npm:@angular/animations@" + angVer + "/bundles/animations.umd.js",
+      "@angular/animations/browser": "npm:@angular/animations@" + angVer + "/bundles/animations-browser.umd.js",
+      "@angular/platform-browser/animations": "npm:@angular/platform-browser@" + angVer + "/bundles/platform-browser-animations.umd.js",
+      "@angular/flex-layout": "npm:@angular/flex-layout@2.0.0-beta.9"
+},
     // packages tells the System loader how to load when no filename and/or no extension
     packages: {
- 
       app: {
         main: './app.module.min.js',
         defaultExtension: 'js'
@@ -72,9 +57,8 @@
 
   // Bootstrap the `AppModule`(skip the `app/main.ts` that normally does this)
   function bootstrap() {
+    //System=require('systemjs');
 
-    // Stub out `app/main.ts` so System.import('app') doesn't fail if called in the index.html
-    //System.set(System.normalizeSync('app/main.ts'), System.newModule({ }));
 
     // bootstrap and launch the app (equivalent to standard main.ts)
     Promise.all([
@@ -88,12 +72,4 @@
     })
     .catch(function(err){ console.error(err); });
   }
-
 })(this);
-
-
-/*
-Copyright 2016 Google Inc. All Rights Reserved.
-Use of this source code is governed by an MIT-style license that
-can be found in the LICENSE file at http://angular.io/license
-*/
